server/requirements.txt
app-version==0.2.1
Django==1.9.2
django-appconf==1.0.1
django-cors-headers==1.1.0
django-extensions==1.6.1
django-filter==0.12.0
django-permission==0.9.2
djangorestframework==3.3.2
djangorestframework-jwt==1.7.2
Markdown==2.6.5
MySQL-python==1.2.5
PyJWT==1.4.0
six==1.10.0
tolerance==0.1.2
Werkzeug==0.11.4
wheel==0.24.0
